requirements.txt
fastapi==0.136.0
uvicorn==0.46.0

SQLAlchemy==2.0.49
python-dotenv==1.2.2
PyYAML==6.0.3

langgraph==1.1.9
langchain-core==1.3.0
langchain-groq==1.1.2

fastembed==0.8.0
faiss-cpu==1.13.2
pinecone>=5.0.0
toons==0.5.4

# Optional database drivers (keep if you use these DB_URI schemes)
psycopg2-binary==2.9.12
PyMySQL==1.1.2
